Add new library fe_power_8k for 8K power functions
