+++ src/petsys_py_lib/daqd.py
@@ -10,7 +10,7 @@ from . import tofpet2c
 from . import info
 from . import bias
 from . import spi
-from . import fe_power
+from . import fe_power, fe_power_8k
 import socket
 from random import randrange
 import struct
@@ -519,11 +519,11 @@ class Connection:
 		for portID, slaveID in self.getActiveFEBDs():
 			if self.__activeUnits[(portID, slaveID)][0] in [ 0x0005 ]:
 				print(f'INFO: Found FEB\D 8K @ ({portID},{slaveID}). Ramping up VDD1 rail after ASIC configuration.')
-				busID_lst = fe_power.detect_active_bus(self, portID, slaveID)
+				busID_lst = fe_power_8k.detect_active_bus(self, portID, slaveID)
 				for busID, moduleVersion in busID_lst:
-					current_dac_setting = fe_power.read_dac(self, portID, slaveID, busID, moduleVersion, 'vdd1')
-					vdd1_iterable = range(current_dac_setting, fe_power.VDD1_PRESET[moduleVersion]["max"] + 1, 2)
-					new_dac_setting = fe_power.ramp_up_rail(self, portID, slaveID, busID, moduleVersion, "vdd1", vdd1_iterable, fe_power.VDD1_PRESET[moduleVersion]["max"], fe_power.VDD1_TARGET)
+					current_dac_setting = fe_power_8k.read_dac(self, portID, slaveID, busID, moduleVersion, 'vdd1')
+					vdd1_iterable = range(current_dac_setting, fe_power_8k.VDD1_PRESET[moduleVersion]["max"] + 1, 2)
+					new_dac_setting = fe_power_8k.ramp_up_rail(self, portID, slaveID, busID, moduleVersion, "vdd1", vdd1_iterable, fe_power_8k.VDD1_PRESET[moduleVersion]["max"], fe_power_8k.VDD1_TARGET)
 					#print(portID, slaveID, busID, moduleVersion, current_dac_setting, new_dac_setting)
 		
 		# Allow some ms for the deserializer to lock to the 8B/10B pattern
